+++ Sources/SBUAvailable.swift
@@ -30,11 +30,9 @@ public class SBUAvailable: NSObject {
 
     private static func isAvailable(key: String) -> Bool {
         guard let appInfo = SBDMain.getAppInfo(),
-              let applicationAttributes = appInfo.applicationAttributes,
-              let premiumFeatureList = appInfo.premiumFeatureList else { return false }
+              let applicationAttributes = appInfo.applicationAttributes else { return false }
         
-        return applicationAttributes.contains(key) ||
-            premiumFeatureList.contains(key)
+        return applicationAttributes.contains(key)
     }